Implemented LFA comments on seperation of concerns in routes
